+++ src/main/java/cart/controller/api/ProductController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.api;
 
 import cart.dto.ProductCreationRequest;
 import cart.dto.ProductModificationRequest;
